fixes #5, allows mysql queries to be when a password is set for the mysql root user.
To use this, log into mysql as root and change the password:
    # mysql -uroot
    mysql> update mysql.user set password=PASSWORD('sekrit') where user='root';
    mysql> flush privileges;

Next, create /root/.my.cnf with the following contents, and chmod 600.
    [mysql]
    password=sekrit

Moonshine recipes will still be able to verify that the user and database exist on deploy. To log in to mysql as root via the @deploy@ system user, use @sudo su -c "mysql@ or @sudo -i@ followed by @mysql@