Fixed an issue with upgrading Passenger Enterprise
The interplay between the package('passenger') and the
package('passenger-enterprise') resources was causing an
issue when upgrading Passenger Enterprise using the new gem
server. By converting the removal of the passenger gem to an
exec and giving the installation of Passenger Enterprise the
title 'passenger', so that other resources (such as building
the passenger apache module) that depend on "package('passenger')"
will execute correctly, the upgrade process now works correctly
with a single deploy.